+++ src/main/java/frc/robot/BuildConstants.java
@@ -1,0 +1,17 @@
+package frc.robot;
+
+/** Automatically generated file containing build version information. */
+public final class BuildConstants {
+    public static final String MAVEN_GROUP = "";
+    public static final String MAVEN_NAME = "reefscape-2025";
+    public static final String VERSION = "unspecified";
+    public static final int GIT_REVISION = 451;
+    public static final String GIT_SHA = "e72a0d42c1f7cecbeb49844abf511c6c84593878";
+    public static final String GIT_DATE = "2025-04-04 12:10:21 EDT";
+    public static final String GIT_BRANCH = "vision-improvements";
+    public static final String BUILD_DATE = "2025-04-04 12:11:36 EDT";
+    public static final long BUILD_UNIX_TIME = 1743783096231L;
+    public static final int DIRTY = 1;
+
+    private BuildConstants() {}
+}
