--- src/main/java/org/littletonrobotics/frc2025/BuildConstants.java
@@ -1,17 +1,0 @@
-package org.littletonrobotics.frc2025;
-
-/** Automatically generated file containing build version information. */
-public final class BuildConstants {
-    public static final String MAVEN_GROUP = "";
-    public static final String MAVEN_NAME = "reefscape-2025";
-    public static final String VERSION = "unspecified";
-    public static final int GIT_REVISION = 445;
-    public static final String GIT_SHA = "fed7d6ccf190012f3b1b5ccf866ac410a7ee0d46";
-    public static final String GIT_DATE = "2025-04-03 21:29:11 EDT";
-    public static final String GIT_BRANCH = "vision-improvements";
-    public static final String BUILD_DATE = "2025-04-03 21:29:14 EDT";
-    public static final long BUILD_UNIX_TIME = 1743730154678L;
-    public static final int DIRTY = 0;
-
-    private BuildConstants() {}
-}
